+++ Nickel.UpdateChecks.UI/ModEntry.cs
@@ -1,21 +1,15 @@
 ﻿using FSPRO;
 using HarmonyLib;
 using Microsoft.Extensions.Logging;
-using Microsoft.Xna.Framework.Graphics;
 using Nanoray.Pintail;
 using Nanoray.PluginManager;
-using Nanoray.Shrike;
-using Nanoray.Shrike.Harmony;
 using Nickel.Common;
 using Nickel.InfoScreens;
 using Nickel.ModSettings;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using System.Reflection;
-using System.Reflection.Emit;
 
 namespace Nickel.UpdateChecks.UI;
 
@@ -104,14 +98,6 @@ public sealed class ModEntry : SimpleMod
 				});
 			};
 		}
-		
-		var harmony = this.Helper.Utilities.Harmony;
-		
-		harmony.Patch(
-			original: AccessTools.DeclaredMethod(typeof(CornerMenu), nameof(CornerMenu.Render))
-			          ?? throw new InvalidOperationException($"Could not patch game methods: missing method `{nameof(CornerMenu)}.{nameof(CornerMenu.Render)}`"),
-			transpiler: new HarmonyMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(CornerMenu_Render_Transpiler))
-		);
 	}
 
 	public override object GetApi(IModManifest requestingMod)
@@ -179,143 +165,5 @@ public sealed class ModEntry : SimpleMod
 		).SubscribeToOnMenuClose(
 			_ => this.UpdateChecksApi.SaveSettings()
 		));
-	}
-
-	[SuppressMessage("ReSharper", "PossibleMultipleEnumeration")]
-	private static IEnumerable<CodeInstruction> CornerMenu_Render_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod)
-	{
-		try
-		{
-			return new SequenceBlockMatcher<CodeInstruction>(instructions)
-				.Find(ILMatches.LdcI4((int)StableSpr.buttons_menu))
-				.Find(ILMatches.Call("Sprite"))
-				.Replace(new CodeInstruction(OpCodes.Call, AccessTools.DeclaredMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(CornerMenu_Render_Transpiler_HijackDraw))))
-				.AllElements();
-		}
-		catch (Exception ex)
-		{
-			Instance.Logger.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, Instance.Package.Manifest.UniqueName, ex);
-			return instructions;
-		}
-	}
-
-	private static void CornerMenu_Render_Transpiler_HijackDraw(Spr? id, double x, double y, bool flipX, bool flipY, double rotation, Vec? originPx, Vec? originRel, Vec? scale, Rect? pixelRect, Color? color, BlendState? blend, SamplerState? samplerState, Effect? effect)
-	{
-		Draw.Sprite(id, x, y, flipX, flipY, rotation, originPx, originRel, scale, pixelRect, color, blend, samplerState, effect);
-
-		var updateSourceMessages = Instance.UpdateChecksApi.UpdateSources
-			.OrderBy(kvp => kvp.Key)
-			.Select(kvp => kvp.Value)
-			.Select(s => Instance.AsUiUpdateSource(s))
-			.OfType<IUiUpdateSource>()
-			.SelectMany(s => s.Messages)
-			.GroupBy(m => m.Level)
-			.ToDictionary(g => g.Key, g => g.ToList());
-
-		var updatesAvailable = Instance.GetPendingModUpdates()
-			.Where(e => Instance.UpdateChecksApi.GetIgnoredUpdateForMod(e.Mod) != e.Version)
-			.ToList();
-
-		List<ISpriteEntry> overlaysToShow = [];
-		var addedTooltips = false;
-
-		if (updatesAvailable.Count > 0)
-			overlaysToShow.Add(Instance.Content.UpdateAvailableOverlayIcon);
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Error, out var messages) && messages.Count > 0)
-			overlaysToShow.Add(Instance.Content.ErrorMessageOverlayIcon);
-		else if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Warning, out messages) && messages.Count > 0)
-			overlaysToShow.Add(Instance.Content.WarningMessageOverlayIcon);
-
-		if (overlaysToShow.Count > 0)
-		{
-			var overlayToShow = overlaysToShow[(int)MG.inst.g.time % overlaysToShow.Count];
-			Draw.Sprite(overlayToShow.Sprite, x, y);
-		}
-
-		if (MG.inst.g.boxes.FirstOrDefault(b => b.key is { } key && key.k == StableUK.corner_mainmenu) is not { } box)
-			return;
-		if (!box.IsHover())
-			return;
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Error, out messages) && messages.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			var i = 0;
-			foreach (var error in messages)
-			{
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::Error{i++}")
-				{
-					Icon = StableSpr.icons_hurt,
-					TitleColor = Colors.textBold,
-					Title = Instance.Localizations.Localize(["settingsTooltip", "error"]),
-					Description = error.Message,
-				});
-			}
-		}
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Warning, out messages) && messages.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			var i = 0;
-			foreach (var error in messages)
-			{
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::Warning{i++}")
-				{
-					Icon = StableSpr.icons_hurtBlockable,
-					TitleColor = Colors.textBold,
-					Title = Instance.Localizations.Localize(["settingsTooltip", "warning"]),
-					Description = error.Message,
-				});
-			}
-		}
-
-		if (updateSourceMessages.TryGetValue(UpdateSourceMessageLevel.Info, out messages) && messages.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			var i = 0;
-			foreach (var error in messages)
-			{
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::Info{i++}")
-				{
-					Icon = StableSpr.icons_hurtBlockable,
-					TitleColor = Colors.textBold,
-					Title = Instance.Localizations.Localize(["settingsTooltip", "info"]),
-					Description = error.Message,
-				});
-			}
-		}
-
-		if (updatesAvailable.Count > 0)
-		{
-			if (!addedTooltips)
-			{
-				// addedTooltips = true;
-				MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new TTDivider());
-			}
-
-			MG.inst.g.tooltips.Add(box.rect.xy + new Vec(15, 15), new GlossaryTooltip($"ui.{Instance.Package.Manifest.UniqueName}::UpdatesAvailable")
-			{
-				Icon = Instance.Content.UpdateAvailableTooltipIcon.Sprite,
-				TitleColor = Colors.textBold,
-				Title = Instance.Localizations.Localize(["settingsTooltip", "updatesAvailableTooltip"]),
-				Description = string.Join("\n", updatesAvailable.Select(e => $"<c=textFaint>{Instance.UpdateChecksApi.GetModNameForUpdatePurposes(e.Mod)}</c> <c=textBold>{e.Mod.Version}</c> -> <c=boldPink>{e.Version}</c>"))
-			});
-		}
 	}
 }